Added an alert dialog when the user clicks on the delete post button on the post menu. The delete button only appears for the current user's posts. Added the function to the repository that queries for the postID, then it saves the imagePath, deletes the posts document, searches for any like relation with the postId, deletes that collection, and then uses the imagePath that was saved to delete the image stored in Firebase. In the FeedViewModel, when the repository function is completed it will remove the post from the list and update the live data, so only that post gets removed and we dont have to refresh the entire feed.
UI changes for consistency:
In addition I implemented a change to the next button in the canvas activity so that the text is grey when not enabled and white when enabled. Also change the color of the loading indicator to tertiary.
Also changed the text fields in the ProfileEditPopup to change colors based on if they are in focus, this logic was done to the @Composable fun EditableTextField in our Components.kt and also changed Biography to Bio in the ProfileEditPopup.